==> R/00_all.qmd <==
---
title: "Project"
format:
  html:
    embed-resources: true
    
project:
  output-dir: ../results
---

# Load Libraries

```{r}
#| message: false
#| warning: false
library("tidyverse")
library("quarto")
```

# 1. Load Data

Below we run the loading script that downloads the raw data and writes `01_dat_load.tsv.`

```{r}
#| echo: true
#| eval: true
quarto_render(input = "01_load.qmd",
              output_file = "01_load.html")

file.rename(from = "01_load.html",
            to = "../results/01_load.html")
```

# 2. Clean data

Cleans the loaded data and writes it to `02_dat_clean.tsv`.

```{r}
#| echo: true
#| eval: true
quarto_render(input = "02_clean.qmd",
              output_file = "02_clean.html")

file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
```

# 3. Augment

Added augmentation the cleaned data and writes it to `03_dat_aug.tsv`.

```{r}
#| echo: true
#| eval: true
quarto_render(input = "03_augment.qmd",
              output_file = "03_augment.html")

file.rename(from = "03_augment.html",
            to = "../results/03_augment.html")
```

# 4. Describe

Describes the cleaned data.

```{r}
#| echo: true
#| eval: true
quarto_render(input = "04_describe.qmd",
              output_file = "04_describe.html")

file.rename(from = "04_describe.html",
            to = "../results/04_describe.html")
```

# 5. Analysis 1

```{r}
#| echo: true
#| eval: true
quarto_render(input = "05_analysis_1.qmd",
              output_file = "05_analysis_1.html")

file.rename(from = "05_analysis_1.html",
            to = "../results/05_analysis_1.html")
```

# 6. Analysis 2

```{r}
#| echo: true
#| eval: true
quarto_render(input = "06_analysis_2.qmd",
              output_file = "06_analysis_2.html")

file.rename(from = "06_analysis_2.html",
            to = "../results/06_analysis_2.html")
```

# 7. Analysis 3

```{r}
#| echo: true
#| eval: true
quarto_render(input = "07_analysis_3.qmd",
              output_file = "07_analysis_3.html")

file.rename(from = "07_analysis_3.html",
            to = "../results/07_analysis_3.html")
```

# 8. Analysis 4

```{r}
#| echo: true
#| eval: true
quarto_render(input = "08_analysis_4.qmd",
              output_file = "08_analysis_4.html")

file.rename(from = "08_analysis_4.html",
            to = "../results/08_analysis_4.html")
```
