==> R/03_augment.qmd <==
---
title: "Augment data"
format:
  html:
    embed-resources: true
---

# Augment Data

## 1. Load libraries

```{r}
#| message: false
#| warning: false
library("tidyverse")
```

## 2. Load data

Loading the data saved in the 02_dat_clean.tsv file

```{r}
genome_metadata_clean <- readr::read_tsv("../data/02_dat_clean.tsv")
```

## 3. Augment data

A new table is created, containing info on the quality of the MAGs. This includes completeness quality, contamination quality as well as overall genome quality.

```{r}
quality_table <- genome_metadata_clean |>
  dplyr::select(Genome, Completeness, Contamination) |>
  dplyr::mutate(
    Completeness_quality = dplyr::case_when(
      Completeness > 90                          ~ "High",
      Completeness >= 50 & Completeness <= 90    ~ "Medium",
      Completeness < 50                          ~ "Low",
      TRUE                                       ~ NA_character_
    ),
    Contamination_quality = dplyr::case_when(
      Contamination < 5    ~ "High",
      Contamination < 10   ~ "Medium",
      Contamination >= 10  ~ "Low",
      TRUE                 ~ NA_character_
    ),
    Overall_quality = dplyr::case_when(
      Completeness > 90  & Contamination < 5    ~ "High",
      Completeness >= 50 & Contamination < 10   ~ "Medium",
      Completeness < 50  & Contamination >= 10  ~ "Low",
      TRUE                                      ~ "Other"
    )
  ) |>
  dplyr::select(
    Genome,
    Completeness_quality,
    Contamination_quality,
    Overall_quality
  )
```

The new data table is then joined with the cleaned data frame using left join.

```{r}
genome_metadata_aug <- genome_metadata_clean |>
  dplyr::left_join(quality_table, by = "Genome")
```

## 4. Annotating MAGs with endometriosis-associated genera

1)  Endobiota Study 2019, "The endobiota study: comparison of vaginal, cervical and gut microbiota between women with stage 3/4 endometriosis and healthy controls" (<https://www.nature.com/articles/s41598-019-39700-6>)

Across these vaginal datasets, the most consistently reported endometriosis-associated genera are:

-   Anaerococcus

-   Ureaplasma

-   Gardnerella

-   Veillonella

-   Corynebacterium

-   Peptoniphilus

-   Candida

-   Alloscardovia

We add a new column to the MAG dataset that flags whether each MAG belongs to a genus repeatedly associated with endometriosis in the vaginal microbiome. Labeling True or False

```{r}
endo_genera <- c(
  "Anaerococcus",
  "Ureaplasma",
  "Gardnerella",
  "Veillonella",
  "Corynebacterium",
  "Peptoniphilus",
  "Candida",
  "Alloscardovia"
)

genome_metadata_aug <- genome_metadata_aug |>
  dplyr::mutate(
    endometriosis_associated = dplyr::case_when(
      Genus %in% endo_genera ~ "Yes",
      TRUE                   ~ "No"
    )
  )

genome_metadata_aug |>
  dplyr::count(endometriosis_associated)
```

## 6. Saving the data

Saving the cleaned data into 03_dat_aug.tsv file located in the data directory.

```{r}
readr::write_tsv(
  x    = genome_metadata_aug,
  file = "../data/03_dat_aug.tsv"
)
```
